+++ src/black/mode.py
@@ -192,6 +192,7 @@ class Preview(Enum):
     fix_power_op_line_length = auto()
     hug_parens_with_braces_and_square_brackets = auto()
     allow_empty_first_line_before_new_block_or_comment = auto()
+    single_line_format_skip_with_multiple_comments = auto()
 
 
 class Deprecated(UserWarning):